backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.10.1
bcrypt==5.0.0
asyncpg==0.31.0
groq==0.11.0
mcp==1.0.0
